First working prototype in TestServer and TestClient in the test package of the api

+++ src/dev/kelvin/api/network/events/IOnConnectionClosed.java
@@ -2,6 +2,6 @@ package dev.kelvin.api.network.events;
 
 public interface IOnConnectionClosed {
 
-    void onConnectionClosed(long uuid);
+    void onConnectionClosed(int uuid);
 
 }
